Initial version of Experiences
* Implemented a temporary Experience class to test out adding and displaying Experiences
* "Add Experience" page includes the ability to upload a custom image
* All current Experiences are shown on the Landing page
* Clicking on any Experience's title leads to its details page

Obviously this is not the final version of Experiences, as I will have to update a lot of stuff after we create the actual database.  This is mainly to get some practice with Flask and plan out how pages will be arranged.

Future plans include displaying the most recent Experiences at the top of the Landing page and showing a user's created Experiences on their Profile page.  I'll have to learn more about SQLAlchemy before I can get those to work.